requirements.txt
ibapi==9.81.1.post1
pandas==2.2.3
numpy==1.26.4
ta==0.11.0
pytz==2024.1
python-dotenv==1.0.1
dash==2.18.1
plotly==5.24.1
dash-bootstrap-components==1.6.0
scipy==1.13.1
matplotlib==3.9.2
requests==2.32.3
